15: Auto focus on the editor when a button for it is clicked.
This prevents some cases when you click on the button and it doesn't
look like anything happens.

Also - resolved an issue with the image insert dialog - the second time it was opened it
would fail to work.

Also - resolved issues related to button handlers with multiple editors on the same page.